Create error module (breaking)
Move previously crate-scope Error, Result, Rv, and
RvError into their own module
