+++ cryptoki/src/functions/decryption.rs
@@ -3,11 +3,10 @@
 //! Decrypting data
 
 use crate::get_pkcs11;
-use crate::Rv;
+use crate::error::{Rv,Result};
 use crate::mechanism::Mechanism;
 use crate::object::ObjectHandle;
 use crate::session::Session;
-use crate::Result;
 use cryptoki_sys::*;
 use std::convert::TryInto;
