+++ cryptoki/src/functions/encryption.rs
@@ -3,11 +3,10 @@
 //! Encrypting data
 
 use crate::get_pkcs11;
-use crate::Rv;
+use crate::error::{Result, Rv};
 use crate::mechanism::Mechanism;
 use crate::object::ObjectHandle;
 use crate::session::Session;
-use crate::Result;
 use cryptoki_sys::*;
 use std::convert::TryInto;
