+++ cryptoki/src/functions/general_purpose.rs
@@ -3,9 +3,8 @@
 //! General-purpose functions
 
 use crate::get_pkcs11;
-use crate::Rv;
+use crate::error::{Result,Rv};
 use crate::context::{Pkcs11,Info,CInitializeArgs};
-use crate::Result;
 use cryptoki_sys::{CK_C_INITIALIZE_ARGS, CK_INFO};
 use std::ptr;
