+++ cryptoki/src/mechanism/elliptic_curve.rs
@@ -1,6 +1,7 @@
 //! ECDH mechanism types
 
-use crate::{types::Ulong, Error, Result};
+use crate::types::Ulong;
+use crate::error::{Error, Result};
 use cryptoki_sys::*;
 use log::error;
 use std::convert::TryFrom;
