+++ cryptoki/src/mechanism/rsa.rs
@@ -4,7 +4,7 @@
 
 use super::{Mechanism, MechanismType};
 use crate::types::Ulong;
-use crate::{Error, Result};
+use crate::error::{Error, Result};
 use cryptoki_sys::*;
 use log::error;
 use std::convert::TryFrom;
