+++ cryptoki/src/slot/mod.rs
@@ -9,7 +9,8 @@
 mod flags;
 pub use flags::*;
 use crate::types::{Ulong, Version};
-use crate::{string_from_blank_padded, Error, Result};
+use crate::error::{Error, Result};
+use crate::string_from_blank_padded;
 use cryptoki_sys::{CK_SLOT_ID, CK_SLOT_INFO, CK_TOKEN_INFO};
 use std::convert::{TryFrom, TryInto};
 use std::fmt::Formatter;
